server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
pydub==0.25.1
moviepy==1.0.3
opencv-python==4.8.0.76
numpy==1.24.3